+++ CHANGELOG.md
@@ -2,8 +2,9 @@
 
 ## v0.2.0-alpha
 
-* Remove engine flag from create integration 
-* Remove is-view from create datastream 
+* Remove engine flag from create integration command
+* Remove is-view from create datastream command
+* Add support for updating credentials used in Snowflake integration
 
 ## v0.1.6-alpha
 * Add Snowflake data stream status support